Add support for DVI monitors
This is a combination patch that merges all the work done to better support
DVI monitors.

Initial patch
Signed-off-by: Robert Winkler <robert.winkler@boundarydevices.com>

dynamic hot-plug interrupt setup (to return original HDMI hot plug
event generation for std HDMI mode, while keeping DVI enhancements
if we work in DVI mode)

Disable interrupts while screen is blanked and in DVI mode. Many
DVI monitors don't properly support hotplug and instead continually
poll the rx signals every few seconds to see if the host is alive.
This will be picked up as a hotplug event and wake everything up.

Handling non-CEA edid parsing and configuration.  This includes
checking for a non-HDMI connection in the audio driver startup
and bailing out.

Default to HDMI mode if we fail to get an edid.

If we don't receive a valid EDID default to HDMI mode.  This is
an HDMI port and selecting DVI while getting a wider range of
resolutions you lose support for audio and cec.  Lets choose
modern and more functional interface as the fallback default.

Co-Authored-By: Matus Kral <matuskral@me.com>
Co-Authored-By: Jon Nettleton <jon@solid-run.com>