tc358743: split set_csi into set_csi and start_csi
The CSI Tx can't be stopped except by pulling the CTx reset.
Split the actual CSI start out of set_csi. This allows to call
it later in s_stream, as the Synopsys Designware MIPI CSI-2 Host
Controller needs to start with the lanes in stop state before
it can sync its PLL to the clock lane.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>