PCI: imx: partially revert 0a057c8b1a7e0
On iMX6dl/q I was seeing hangs occasionally when no pcie card
was connected to the board.  I traced this down to disabling
the pcie clock introduced in this commit.

commit 0a057c8b1a7e0bce50d51ed37a1603f23dc2efd9
Author: Richard Zhu <Richard.Zhu@freescale.com>
Date:   Thu Dec 25 14:36:44 2014 +0800

    MLK-10058-4 pci: imx6: refine imx6sx pcie pm

Since this seems to only be tested with the iMX6SX and I see no
benefit on the other iMX6 chipsets since the PCI bus is clocked
directly from the AXI clock, I have wrapped this in the XTEME
power savings config option as well.

I also tested on our platform and saw no power differences if the
clocks were enabled or disabled with no card present.