dts: hb/hb2: Add basic support for the lvds port
This adds very basic support for the single channel lvds port
that is on the end of the Hummingboard and Hummingboard edge.
It is disabled by default because it can not be hotplug detected
and we don't need memory reserved for something that isn't
going to be used by many people.

To enable it you need to enable the second framebuffer device and
then enable the ldb.  You should also take note to change the fb
aliases and the crtc depending on the SOC variant you are using
and whether this should be the primary framebuffer device or not.
By default I have it aliased to be /dev/fb2.