dts: hummingboard: Move gpio controlled regulator to card reset
Some sdhc cards would hard lock in the firmware because on shutdown
the regulator would be turned off, however as soon as the bootloader
firmware would start it would be turned on.  This time period
was short enough that it would crash some cards firmware.  Instead
we will set up that regulator as a reset.  This allows both voltage
transitioning and fallback for 1.8V cards to work, as well as
consistent shutdowns.