mmc: core: Don't fail UHS transition if host max_current fails.
During the transition to UHS we query the max_current of the host
and then attempt to negotiate that setting.  If it fails then
we exit the entire UHS transition.  Since this is the max_current
of the host and not the card try lower current settings until
they are accepted and we can move forward with the transition.

This should be safe since we are going from max_current down, and
any problems in the signalling should be caught when the voltage
switch happens.