gpu: galcore: Clean up spurious errors with gpu idling.
The GPU driver will report errors in the logs if the following
conditions are met, the Event submission finishes and the
gpu hardware is idle, and the pm sees that we are within an
interrupt context.  We can avoid this and also skip a lot of
needless locking by just tracking where the Event submission has
come from and then skip idling the GPU.  Basically if it is a
command submission we know the hardware will be doing something
soon so there is no reason to try to do an idle suspend.