crypto: hash - Increase digest size check
Now that we are setting a statesize we find out that our digests
are considerably larger than expected by existing checks in place.
This check was added with upstream commit
ca786dc738f4f583b57b1bba7a335b5e8233f4b0 and seemed to be an
arbitrary choice to fix another problem.  The digest has to fit
on the stack which is 2 pages.  PAGE_SIZE / 2 seems to be pushing
the limit but is not causing any faults so I will continue to use
it.  Ultimately I think the caam driver needs to be reworked to
handle context's differently.