crypto: fix libcrc32c loading
The backport commit of d550448f540a9addd7115c419dd9044c96890569
caused a partial breakage in loading the libcrc32c module.  This
caused breakage in filesystem and modules that depending on it.

By reverting the changes to shash.c and keeping the changes to
ahash.c we retain the intended checks again crypto modules but
allow to function libcrc32c properly.